tests: Added unit test for test_mode function
